Moved provider to lib
Code cleanup
